requirements.txt
Django==4.1.5
djangorestframework==3.14.0
beautifulsoup4==4.11.1
requests==2.28.1
